Refactor image and album handling in modules
Significant refactor of the images and albums feature in the modules. The changes shift core image handling functionality to the album controller and improve error handling in the album service. The image handling process is also optimized in the image service.

Changes:
- src/modules/images/controllers/image.controller.ts:
  + Deleted the image controller to shift functionality to album controller.
- src/modules/images/services/image.service.ts:
  + Replaced CreateImageDto with direct albumId.
  + Optimized the image creation process.
  + Simplified the process to find all images by albumId.
- src/modules/images/controllers/album.controller.ts:
  + Moved and renamed from src/modules/images/controllers/album.controller.ts.
  + Improved and extended the capability of album controller.
- src/modules/images/services/album.service.ts:
  + Improved error handling with NotFoundException.
  + Optimized the thumbnail generation process.
- src/app.module.ts:
  + Updated imports to accommodate shift in functionality.
- src/modules/firebase/services/storage.service.ts:
  + Adjusted return type comment in uploadImage method.
- src/modules/images/entities/image.entity.ts:
  + Removed contentType property.
- src/modules/images/images.module.ts:
  + Renamed to album.module.ts.
  + Removed image controller import and provider.

+++ src/modules/images/services/album.service.ts
@@ -1,15 +1,16 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 
 import { InjectRepository }                from '@mikro-orm/nestjs';
 import { EntityRepository, QBFilterQuery } from '@mikro-orm/core';
+import * as path                           from 'node:path';
 import { v4 }                              from 'uuid';
 
 import { CommonService }                          from '@common/common.service';
 import { generateImageObject, generateThumbnail } from '@common/utils/file.utils';
 import { StorageService }                         from '@modules/firebase/services/storage.service';
+import { QueryAlbumDto }                          from '@modules/images/dtos/query-album.dto';
 import { CreateAlbumDto }                         from '../dtos/create-album.dto';
 import { AlbumEntity }                            from '../entities/album.entity';
-import { QueryAlbumDto }                          from '@modules/images/dtos/query-album.dto';
 
 @Injectable()
 export class AlbumService {
@@ -27,9 +28,9 @@ export class AlbumService {
   ): Promise<AlbumEntity> {
     await this._checkAlbumExists(createAlbumDto.name);
 
-    const albumId = v4();
-    const basePath = `companies/${ companyId }/media/albums/${ albumId }`;
-    const album = this.albumRepository.create({
+    const albumId: string = v4();
+    const basePath: string = `companies/${ companyId }/media/albums/${ albumId }`;
+    const album: AlbumEntity = this.albumRepository.create({
       ...createAlbumDto,
       id: albumId,
       company: companyId,
@@ -37,7 +38,7 @@ export class AlbumService {
     });
 
     if (cover) {
-      cover.originalname = 'cover.webp';
+      cover.originalname = 'cover' + path.extname(cover.originalname);
       const {filepath, fileUrl} = await this._storageService.uploadImage(basePath, cover, true);
       album.cover = generateImageObject(cover, filepath, fileUrl);
 
@@ -52,7 +53,7 @@ export class AlbumService {
       const {
         filepath: thumbFilepath,
         fileUrl: thumbFileUrl
-      } = await this._storageService.uploadImage(basePath, coverThumbnailFile);
+      } = await this._storageService.uploadImage(basePath, coverThumbnailFile, true);
 
       album.coverThumbnail = generateImageObject(coverThumbnailFile, thumbFilepath, thumbFileUrl);
     }
@@ -76,12 +77,19 @@ export class AlbumService {
     return this.albumRepository.findAll({where: whereFilter});
   }
 
-  async findOne(id: string): Promise<AlbumEntity> {
-    return this.albumRepository.findOneOrFail({id});
+  async findOne(id: string, companyId: string): Promise<AlbumEntity> {
+    const album = await this.albumRepository.findOne({id, company: {id: companyId}});
+
+    if (!album) throw new NotFoundException('ALBUM_NOT_FOUND');
+
+    return album;
   }
 
   async remove(id: string): Promise<void> {
-    const album = await this.albumRepository.findOneOrFail({id});
+    const album = await this.albumRepository.findOne({id});
+
+    if (!album) throw new NotFoundException('ALBUM_NOT_FOUND');
+
     await this._commonService.removeEntity(album);
   }
